requirements.txt
#ollama 
sentence-transformers
faiss-cpu
datasets
transformers
langchain-huggingface 
langchain_community 
huggingface_hub
langchain
sentence-transformers
fastapi[standard]
#streamlit
psycopg2-binary
jproperties
python-dotenv
dynaconf